requirements.txt
aiogram
python-dotenv
databases[postgresql]
asyncpg